IMPALA-3105: avoid overrunning allocated tuple buffer
This change reworks allocation of tuple buffers. Previously the code
tried to estimate via MaxTupleBufferSize() the maximum size of buffer it
would need to allocate, assuming that AT_CAPACITY_MEM_USAGE was not
exceeded. However, AT_CAPACITY_MEM_USAGE is a soft limit so can be
exceeded.

The fix is to change the code to set the batch's 'capacity_', which is a
hard limit, to a value such that FIXED_LEN_BUFFER_LIMIT <=
AT_CAPACITY_MEM_USAGE won't be exceeded.

Change-Id: Idfd9cd681875821c1c379d97586d3f4850aae622
Reviewed-on: http://gerrit.cloudera.org:8080/2473
Reviewed-by: Tim Armstrong <tarmstrong@cloudera.com>
Tested-by: Internal Jenkins