kudu: add a flag for scanner timeout and increase default
The default Kudu scanner RPC timeout is 15 seconds. This is a bit short
for concurrent workloads: when I looped TPC-H queries on a small cluster
for a few days, I saw one failure due to a timeout. When I dug into the
logs, it seems that nothing went wrong -- just the disks were slow to
respond at one point.

This bumps the default to 60 seconds and makes it configurable in case
we have any issues in production installs where we see timeouts even
with the higher default.

Change-Id: I6d345cdc78f095fd69624cc854ec46843db4158b
Reviewed-on: http://gerrit.cloudera.org:8080/2705
Reviewed-by: Casey Ching <casey@cloudera.com>
Tested-by: Internal Jenkins