IMPALA-3302: Fix ASAN use-after-free failure in Kudu scanner
When fetching VARCHAR data from Kudu, Impala was using a pointer that
was owned and freed by the Kudu client. There was already a mechanism to
copy string data but it was only applied to STRING columns. The fix is
to apply the same mechanism to VARCHAR columns. That's done by adding
the VARCHAR column to the list of string type columns/slots that need to
be copied.

Change-Id: Ib513aae531086355e77ac9c35ed38728efaa0be9
Reviewed-on: http://gerrit.cloudera.org:8080/2749
Reviewed-by: Dan Hecht <dhecht@cloudera.com>
Tested-by: Internal Jenkins