Add NumFunctions and NumInstructions counters to LlvmCodeGen
These counters give a rough idea of how much "work" the module
optimization and compilation need to do, and are especially useful
when reasoning about the cost (or benefit) of new codegen
changes. This change doesn't have any measurable impact on codegen
time for a reasonably-sized module (372 functions, ~5k instructions).

Change-Id: I9a7f4aa02239d0738d891b09fa4531b560428898
Reviewed-on: http://gerrit.cloudera.org:8080/2756
Reviewed-by: Dan Hecht <dhecht@cloudera.com>
Tested-by: Internal Jenkins