IMPALA-3326: TestShowCreateTable switched to use unique_database
There were two tests that use the same database with common setup/
teardown methods. If they were to run in parallel and one finished
before the other, the database would be removed while a test is
running, causing it to fail. Likewise if one started after the
other began.

The fix uses the unique_database fixture in each test so the tests
will not interfere with each other.

Change-Id: I4855a3a6ff0cea9f6cb3f8433e8705762afefe7d
Reviewed-on: http://gerrit.cloudera.org:8080/2768
Reviewed-by: Michael Brown <mikeb@cloudera.com>
Tested-by: Internal Jenkins