IMPALA-1920: Fix userinfo and authority parsing.
IMPALA-1170 fixed parsing of host parts in urls under the presence of
multiple '@' characters. However the parser contained another issue
where userinfo was erroneously extracted if the query part contained
an '@' but no userinfo was present. This change makes the parser only
extract userinfo before the first '/'.

This change also fixes an issue where the authority part was not
correctly parsed in absence of a terminating slash.

Change-Id: I6486aecc7344f40d6f85d70960866fbb45bea2e1
Reviewed-on: http://gerrit.cloudera.org:8080/2777
Reviewed-by: Lars Volker <lv@cloudera.com>
Tested-by: Internal Jenkins