deploy: move cluster argument back into main parser
It was confusing for --cluster to be a parent parser argument. It meant that
the following command line doesn't work:

  deploy.py --host foo --cluster c1 clone IMPALA-KUDU IMPALA-1

Instead, users had to do the following:

  deploy.py --host foo clone --cluster c1 IMPALA-KUDU IMPALA-1

Which is unintuitive; most people expect optional parameters to be grouped
together. So let's move the --cluster argument out of the parent parser and
into the main parser. That means it must be provided in the same group as
--host, --user, and --password.

Change-Id: I92dc5d22328eb12951d086a5070aec11a0e5c443
Reviewed-on: http://gerrit.cloudera.org:8080/2837
Reviewed-by: Casey Ching <casey@cloudera.com>
Tested-by: Casey Ching <casey@cloudera.com>