IMPALA-3418: The Impala FE project relies on Z-tools snapshot builds
The main motivation is that the new build will not allow builds to be
dependent on non build snapshots. The other motivation is that through
dependency domination pull in potential updates of java-cup-runtime.

The change replaces the java-cup version 0.11-a-czt02-20150912.002524-39
with 0.11-a-czt02-cdh. And cup-maven plugin 1.6-20150912.002559-50 with
1.6-cdh.

These versions along with the java-cup-runtime that is a dependency for both
has been deployed to our external maven repo.

A word of caution:
This dependency (and probably others), do not work with any recent version of
maven. Our internal toolchain version 3.0.4 released in 2012 is compatible with
this version but an attempt of building with a more recent will fail because of
internal incompatibilities.

Change-Id: I1540f5b6d207c573a60d05eecd91c778c5401261
Reviewed-on: http://gerrit.cloudera.org:8080/2859
Reviewed-by: Alex Behm <alex.behm@cloudera.com>
Tested-by: Alex Behm <alex.behm@cloudera.com>