IMPALA-3383: don't dump core when misconfigured.
Calling LOG(FATAL) calls abort(), which will generate a core dump in
many cases. This is sometimes useful, but sometimes not (e.g. if we were
just passed a bad command-line argument, we know what the problem is
from the log message).

This patch adds a CLEAN_EXIT_WITH_ERROR macro that logs to the ERROR
log, flushes it, and then exits the process with exit code 1. This
prevents custom cluster tests from generating unnecessary core dumps.

Change-Id: If1686bf72df9420512e14af1a55279a064b633c9
Reviewed-on: http://gerrit.cloudera.org:8080/2862
Reviewed-by: Tim Armstrong <tarmstrong@cloudera.com>
Tested-by: Internal Jenkins