IMPALA-3412: fix CHAR codegen crash in tuple comparator
Attempting to codegen a sort where the sort expr has a CHAR type as an
intermediate result fails completely.

The problem is that ScalarFnCall checked whether its input arguments
were CHAR to disable codegen, but didn't check its output.

This patch also replaces some incorrect codegen CHAR logic that should
not be executed with DCHECKs.

Testing:
The test is a minimal reproduction of the issue. The test is executed
both by the sorter and top-n nodes so covers both cases.

Change-Id: I189073d46a10988803d572928a38f4a718690fa3
Reviewed-on: http://gerrit.cloudera.org:8080/2876
Reviewed-by: Tim Armstrong <tarmstrong@cloudera.com>
Tested-by: Internal Jenkins