IMPALA-3502: Fix race in the coordinator while updating filter routing table
This commit fixes an issue where a fragment may send an
UpdateFilter message to the coordinator while the latter is still
updating the runtime filter routing table. The fix is to decouple the
update of the filter routing table from starting the fragments. With
this fix, the coordinator will finish populating the filter routing
table before it starts any remote fragments.

Change-Id: Iecc737106fd38aa4af0c72959a577adfb413728d
Reviewed-on: http://gerrit.cloudera.org:8080/3018
Reviewed-by: Dimitris Tsirogiannis <dtsirogiannis@cloudera.com>
Tested-by: Internal Jenkins